xtables: use exponential backoff when waiting for xtables lock to be free
Currently, if the -w option is used, iptables (and related commands)
will wait for the xtables lock to be freed. This is done by polling
the lock with 1 second intervals. Often we have seen that the lock
is held much less than that resulting in unnecessary delay when
trying to acquire the lock.

Introduce an exponential delay starting from 10ms. If the lock cannot
be acquired after 10ms, increase the delay to 20ms, then 40ms, etc.
The starting value of 10ms is chosen based on obseravations, there may be
a better value.

CRs-Fixed: 982615
Change-Id: I8541d792e03d0715541af46a8107fe3f0ab874e4